fix error when cropping small images

+++ src/marker_utils.rs
@@ -61,8 +61,8 @@ pub mod marker_processing {
     pub fn resize_if_larger_dims(img: Mat, out_size: &Size) -> Result<Mat, opencv::Error> {
         // Resize l'image si elle est plus petite que les dimensions d'output
         let img_size = img.size()?;
-        let width_ratio: f64 = ((out_size.width / img_size.width) as f64).max(1.);
-        let height_ratio: f64 = ((out_size.height / img_size.height) as f64).max(1.);
+        let width_ratio: f64 = (out_size.width as f64 / img_size.width as f64).max(1.);
+        let height_ratio: f64 = (out_size.height as f64 / img_size.height as f64).max(1.);
 
         if width_ratio == 1. && height_ratio == 1. {
             return Ok(img);
